+++ dinky-admin/src/main/java/org/dinky/controller/TaskController.java
@@ -34,6 +34,7 @@ import org.dinky.data.enums.ProcessType;
 import org.dinky.data.enums.Status;
 import org.dinky.data.exception.NotSupportExplainExcepition;
 import org.dinky.data.exception.SqlExplainExcepition;
+import org.dinky.data.model.JarSubmitParam;
 import org.dinky.data.model.Task;
 import org.dinky.data.result.ProTableResult;
 import org.dinky.data.result.Result;
@@ -45,8 +46,6 @@ import org.dinky.job.JobResult;
 import org.dinky.mybatis.annotation.Save;
 import org.dinky.service.TaskService;
 import org.dinky.trans.ExecuteJarParseStrategyUtil;
-import org.dinky.trans.dml.ExecuteJarOperation;
-import org.dinky.trans.parse.ExecuteJarParseStrategy;
 import org.dinky.utils.SqlUtil;
 
 import java.util.Arrays;
@@ -307,7 +306,7 @@ public class TaskController {
         String sqlStatement = taskDTO.getStatement();
         String[] statements = SqlUtil.getStatements(sqlStatement);
         FlinkJarSqlConvertVO flinkJarSqlConvertVO = new FlinkJarSqlConvertVO();
-        flinkJarSqlConvertVO.setJarSubmitParam(ExecuteJarOperation.JarSubmitParam.empty());
+        flinkJarSqlConvertVO.setJarSubmitParam(JarSubmitParam.empty());
         if (ArrayUtil.isEmpty(statements)) {
             flinkJarSqlConvertVO.setInitSqlStatement(sqlStatement);
             return Result.succeed(flinkJarSqlConvertVO);
@@ -322,7 +321,7 @@ public class TaskController {
             return Result.succeed(flinkJarSqlConvertVO);
         }
         String lastSqlStatement = statements[lastExecuteJarSqlStatementIndex];
-        ExecuteJarOperation.JarSubmitParam info = ExecuteJarParseStrategy.getInfo(lastSqlStatement);
+        JarSubmitParam info = JarSubmitParam.getInfo(lastSqlStatement);
         flinkJarSqlConvertVO.setJarSubmitParam(info);
         String sql = Arrays.stream(ArrayUtil.remove(statements, lastExecuteJarSqlStatementIndex))
                 .map(x -> x + ";")
@@ -334,7 +333,7 @@ public class TaskController {
     @PostMapping("/flinkJarFormConvertSql")
     @ApiOperation("FlinkJar FormConvertSql")
     public Result<String> flinkJarFormConvertSql(@RequestBody FlinkJarSqlConvertVO dto) {
-        ExecuteJarOperation.JarSubmitParam jarSubmitParam = dto.getJarSubmitParam();
+        JarSubmitParam jarSubmitParam = dto.getJarSubmitParam();
         Dict objectMap = Dict.create()
                 .set("uri", Opt.ofNullable(jarSubmitParam.getUri()).orElse(""))
                 .set(
