+++ dinky-client/dinky-client-base/src/main/java/org/dinky/trans/parse/ExecuteJarParseStrategy.java
@@ -19,21 +19,12 @@
 
 package org.dinky.trans.parse;
 
-import org.dinky.parser.SqlSegment;
 import org.dinky.trans.dml.ExecuteJarOperation;
 
 import org.apache.flink.table.operations.Operation;
 import org.apache.flink.table.planner.parse.AbstractRegexParseStrategy;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import cn.hutool.core.bean.BeanUtil;
-import cn.hutool.core.bean.copier.CopyOptions;
-import cn.hutool.core.util.StrUtil;
 
 public class ExecuteJarParseStrategy extends AbstractRegexParseStrategy {
     private static final String PATTERN_STR = "^EXECUTE\\s+JAR\\s+WITH\\s*\\(.+\\)";
@@ -42,30 +33,6 @@ public class ExecuteJarParseStrategy extends AbstractRegexParseStrategy {
 
     public ExecuteJarParseStrategy() {
         super(PATTERN);
-    }
-
-    public static ExecuteJarOperation.JarSubmitParam getInfo(String statement) {
-        statement = statement.replace("\r\n", " ").replace("\n", " ") + " ENDOFSQL";
-        SqlSegment sqlSegment = new SqlSegment("with", "(with\\s+\\()(.+)(\\))", "',");
-        sqlSegment.parse(statement);
-        List<String> bodyPieces = sqlSegment.getBodyPieces();
-        Map<String, String> keyValue = getKeyValue(bodyPieces);
-        return BeanUtil.toBean(
-                keyValue,
-                ExecuteJarOperation.JarSubmitParam.class,
-                CopyOptions.create().setFieldNameEditor(s -> StrUtil.toCamelCase(s, '-')));
-    }
-
-    private static Map<String, String> getKeyValue(List<String> list) {
-        Map<String, String> map = new HashMap<>();
-        Pattern p = Pattern.compile("'(.*?)'\\s*=\\s*'(.*?)'");
-        for (String s : list) {
-            Matcher m = p.matcher(s + "'");
-            if (m.find()) {
-                map.put(m.group(1), m.group(2));
-            }
-        }
-        return map;
     }
 
     @Override
