+++ dinky-core/src/main/java/org/dinky/job/runner/JobJarRunner.java
@@ -24,6 +24,7 @@ import org.dinky.classloader.DinkyClassLoader;
 import org.dinky.data.exception.DinkyException;
 import org.dinky.data.job.JobStatement;
 import org.dinky.data.job.SqlType;
+import org.dinky.data.model.JarSubmitParam;
 import org.dinky.data.result.InsertResult;
 import org.dinky.data.result.SqlExplainResult;
 import org.dinky.gateway.Gateway;
@@ -222,7 +223,7 @@ public class JobJarRunner extends AbstractJobRunner {
         for (String sql : statements) {
             String sqlStatement = jobManager.getExecutor().pretreatStatement(sql);
             if (ExecuteJarParseStrategy.INSTANCE.match(sqlStatement)) {
-                uriList.add(ExecuteJarParseStrategy.getInfo(statement).getUri());
+                uriList.add(JarSubmitParam.getInfo(statement).getUri());
                 break;
             }
         }
